move the base class for auxiliary modules to ewoms/disc/common
that's where it belonged in the first place.
