+++ ewoms/disc/common/baseauxiliarymodule.hh
