+++ ewoms/disc/common/fvbasediscretization.hh
@@ -45,12 +45,12 @@
 #include "fvbaseprimaryvariables.hh"
 #include "fvbaseintensivequantities.hh"
 #include "fvbaseextensivequantities.hh"
+#include "baseauxiliarymodule.hh"
 
 #include <ewoms/parallel/gridcommhandles.hh>
 #include <ewoms/parallel/threadmanager.hh>
 #include <ewoms/linear/nullborderlistmanager.hh>
 #include <ewoms/common/simulator.hh>
-#include <ewoms/aux/baseauxiliarymodule.hh>
 #include <ewoms/common/alignedallocator.hh>
 #include <ewoms/common/timer.hh>
 #include <ewoms/common/timerguard.hh>
